+++ CHANGES.md
@@ -8,3 +8,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Review tool: mouse-wheel scroll now lets you seek in video. Up - previous 
   frame, Down - next frame, combined with Shift modifier- jump 10 frames back 
   or forward.
+## [0.1.0-13] - 2021-06-10
+### Changed
+- Review tool: mouse wheel scroll modifier Shift for single frame seek
+  and Ctrl + Shift for jumps as just scroll interferes with scrollbar
+  functionality.
+